Remove bfuncs from cb_summary_stats_few_cats
Discussed in #3

The use of bfuncs to calculate categorical statistics is deprecated. The freq_table() function is now part of the freqtables package. However, we don't really need to use freqtables to calculate the n, cumulative n, and percentage. I'm just using standard dplyr functions to do these calculations now.