Update to rlang 0.4.0
Discussed in #3

In the process of trying to make this change, it turns out that rlang::sym(.x) is no longer valid code. It returns the following error: Error in ``rlang::sym()`` at codebookr/R/cb_summary_stats_many_cats.R:19:2: Can't convert a function to a symbol.. As of rlang 0.4.0, this line of code is unnecessary. I'm removing it from the code and replacing all of the !!xs with {{ .x }} -- the new preferred tidy evaluation syntax.