Cause cb_summary_stats.R to ignore `haven_labelled` and `vctrs_vctr`
Discussed in #3 

When you import data from SAS, Stata, or SPSS using Haven, it adds two classes to variables with value labels:  `haven_labelled` and `vctrs_vctr`. Passing these columns to `codebook()` results in an error. One way to get around this is simply to set the `col_type` attribute. However, because Haven labeled data is so common, we decided to specifically look for and remove those classes in this section of code that is trying to determine the column type. It should not remove those class from the column generally.