Update cb_get_col_attriubutes.R to use Haven labels
Discussed in #3

When we import data from Stata, SAS, or SPSS with labels, the attributes are called $label for variable labels and $labels for value labels. Currently, codebook() cannot automatically make use of those attributes because it only recognizes the attributes description, source, and col_type. It's relatively easy to manually set the value of the description attribute to the value of the label attribute. However, because Haven labeled data is so common, we decided to specifically look for $label and $labels in cb_get_col_attibutes.