Fix `Value = date` error in `cb_summary_stats_time.R`
Discussed in #9

The section of code that calculates the mode value still had the date column hardcoded into it from when I originally created and test the function. This explains why the date column had a mode value and didn't throw an error, but other columns with mode value did.