Update README
I received an email that said, "Today I found the 'Codebookr' package and was amazed by how easily it turns variable names, labels and values into a very neat looking codebook. However, I am working with quite sensitive data and want to omit all summary stats, and not just of specific columns, as is the case with the argument 'no_summary_stats' when creating a codebook. At least, that’s how far I understood. I tried passing `colnames(df)` into this argument to eliminate the summary statistics for all columns, however without success. Do you have any advice on how to omit the summary statistics in a codebook for all columns?"

So, I added a new example to README that shows how to omit the summary statistics for all of the columns in the data frame being passed to codebook()